Split REPL and MCP manager god objects into focused modules
Extract REPL prompt activation and slash-command routing into dedicated helpers to shrink repl.ts and isolate control flow.

Replace McpManager internals with resolver/registry/validation/auth flows and keep a thin facade for existing session integration points.
